[refactor]: 프로필/사이드바 컴포넌트 위치 변경(#164)
- 기존 : giftBox 내부에서 프로필과 사이드바가 구현되어 있음
- 변경 : 사이드바는 ui 컴포넌트로, 프로필은 feature 컴포넌트로 이동
