+++ src/layouts/MyPage/GiftBox/index.tsx
@@ -1,0 +1,5 @@
+const GiftBox = () => {
+  return <> </>;
+};
+
+export default GiftBox;
